Crate midir

Modules

os

Structs

ConnectError

An error that can occur when trying to connect to a port.

InitError

An error that can occur during initialization (i.e., while creating a MidiInput or MidiOutput object).

MidiInput

An instance of MidiInput is required for anything related to MIDI input. Create one with MidiInput::new.

MidiInputConnection

Represents an open connection to a MIDI input port.

MidiInputPort

An object representing a single input port. How the port is identified internally is backend-dependent. If the backend allows it, port objects remain valid when other ports in the system change (i.e. it is not just an index).

MidiOutput

An instance of MidiOutput is required for anything related to MIDI output. Create one with MidiOutput::new.

MidiOutputConnection

Represents an open connection to a MIDI output port.

MidiOutputPort

An object representing a single output port. How the port is identified internally is backend-dependent. If the backend allows it, port objects remain valid when other ports in the system change (i.e. it is not just an index).

Enums

ConnectErrorKind

The kind of error for a ConnectError.

Ignore

An enum that is used to specify what kind of MIDI messages should be ignored when receiving messages.

PortInfoError

An error that can occur when retrieving information about available ports.

SendError

An error that can occur when sending MIDI messages.

Traits

MidiIO

Trait that abstracts over input and output ports.

Type Aliases

MidiInputPorts

A collection of input ports.

MidiOutputPorts

A collection of output ports.